Tratamento de erros 404

+++ projeto-contatos/src/main/java/br/senac/tads/dsw/projetocontatos/ContatoRestController.java
@@ -4,10 +4,13 @@
  */
 package br.senac.tads.dsw.projetocontatos;
 
+import java.net.URI;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +20,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
 @RequestMapping("/contatos")
@@ -32,19 +37,48 @@ public class ContatoRestController {
     
     @GetMapping("/consulta/{id}")
     public Contato consultar(@PathVariable("id") int id) {
-        return service.findById(id);
+//        Contato c = service.findById(id);
+//        if (c == null) {
+//            // Contato não encontrado
+//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, 
+//                    "Contato com ID " + id + " não encontrado");
+//        }
+//        return c;
+//        // Versão com optional
+//        Optional<Contato> optContato = service.findByIdComOptional(id);
+//        if (optContato.isEmpty()) {
+//             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+//                     "Contato com ID " + id + " não encontrado");
+//        }
+//        return optContato.get();
+        
+        // Versão com optional funcional
+        return service.findByIdComOptional(id).orElseThrow(() -> 
+                new ResponseStatusException(HttpStatus.NOT_FOUND,
+                     "Contato com ID " + id + " não encontrado"));
     }
-    
     
     @PostMapping("/incluir")
     public ResponseEntity<?> incluir(@RequestBody Contato c) {
         service.save(c);
-        return ResponseEntity.ok().build();
+//        return ResponseEntity.created(
+//                URI.create("http://localhost:8080/contatos/consultar/" 
+//                        + c.getId())).build();
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentContextPath()
+                .path("/contatos/consultar/{id}")
+                .buildAndExpand(c.getId()).toUri();
+        return ResponseEntity.created(location).build();
     }
     
     @PutMapping("/alterar/{id}")
     public ResponseEntity<?> alterar(@PathVariable("id") int id,
             @RequestBody Contato c) {
+        Optional<Contato> optContato = service.findByIdComOptional(id);
+        if (optContato.isEmpty()) {
+             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                     "Contato com ID " + id + " não encontrado");
+        }
         c.setId(id); // Ignora o ID informado no JSON - ID da URL tem predominancia
         service.save(c);
         return ResponseEntity.ok().build();
@@ -52,6 +86,11 @@ public class ContatoRestController {
     
     @DeleteMapping("/excluir/{id}")
     public ResponseEntity<?> excluir(@PathVariable("id") int id) {
+        Optional<Contato> optContato = service.findByIdComOptional(id);
+        if (optContato.isEmpty()) {
+             throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                     "Contato com ID " + id + " não encontrado");
+        }
         service.deleteById(id);
         return ResponseEntity.ok().build();
     }
